+++ designer/client/src/components/graph/Graph.tsx
@@ -16,7 +16,7 @@ import { isTouchEvent, LONG_PRESS_TIME } from "../../helpers/detectDevice";
 import { NotificationActions } from "../../http/HttpService";
 import { batchGroupBy } from "../../reducers/graph/batchGroupBy";
 import { prepareNewNodesWithLayout } from "../../reducers/graph/utils";
-import { FRAGMENT_TEMPLATE_ID } from "../../reducers/selectors/componentGroups";
+import { FRAGMENT_TEMPLATE_ID } from "../../reducers/selectors/appendFragmentCreator";
 import { UserSettings } from "../../reducers/userSettings";
 import { Edge, NodeId, NodeType, ProcessDefinitionData, ScenarioGraph } from "../../types";
 import { StickyNoteType } from "../../types/stickyNote";
