+++ designer/client/src/components/graph/node-modal/isAggregate.tsx
@@ -1,0 +1,5 @@
+import type { NodeType } from "../../../types";
+
+export function isAggregate(node: NodeType) {
+    return ["aggregate-session", "aggregate-sliding", "aggregate-tumbling"].includes(node.nodeType);
+}
